+++ app/src/main/java/org/android/go/sopt/ui/main/home/HomeFragment.kt
@@ -8,8 +8,9 @@ import org.android.go.sopt.binding.BindingFragment
 import org.android.go.sopt.databinding.FragmentHomeBinding
 import org.android.go.sopt.ui.main.home.adapter.MultiViewAdapter
 import org.android.go.sopt.ui.main.home.adapter.MultiViewItem
-import org.android.go.sopt.ui.main.home.data.DataSource
+import org.android.go.sopt.ui.main.home.adapter.data.DataSource
 
+/** 멀티 뷰 타입의 리사이클러뷰 적용 (텍스트 타입, 이미지 타입) */
 class HomeFragment : BindingFragment<FragmentHomeBinding>(R.layout.fragment_home) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
