+++ app/src/main/java/org/android/go/sopt/ui/main/search/SearchFragment.kt
@@ -6,6 +6,7 @@ import org.android.go.sopt.R
 import org.android.go.sopt.binding.BindingFragment
 import org.android.go.sopt.databinding.FragmentSearchBinding
 
+/** 리사이클러뷰 Selection 라이브러리 사용해보기 */
 class SearchFragment : BindingFragment<FragmentSearchBinding>(R.layout.fragment_search) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
